requirements.txt
streamlit>=1.33
pypdf>=4.0
openai>=1.40
requests>=2.31
beautifulsoup4>=4.12
python-dotenv>=1.0
